Rebuild only the failed builds
Homu now remembers the previous build results, which allows it to retry
only the failed builds next time. This change is only applied to the
repositories using Buildbot.

This commit also adds `clean` command, which cleans up the build
results. This can be used to force Homu to build the pull request from
the beginning.

Note that the build results are reusable only when no other commits were
merged into `master`.

Fixes #56.

[breaking-change]