Retrieve head SHA manually if it is invalid
Sometimes, GitHub sends the head SHA of a PR as 0000000 through the
webhook. This is called a "null commit", and seems to happen when GitHub
internally encounters a race condition. Last time, it happened when
squashing commits in a PR. In this case, we just try to retrieve the
head SHA manually.

Fixes #99.