Don't report stdout when a Git command fails
For some reason, `subprocess.check_output` is blocked indefinitely
sometimes, leaving behind a zombie process. It seems that the
`self.stdout.read()` call in `Popen.communicate` tries to read from the
child process, but as the child process is already gone, it is blocked.
Until a true reason is found, just turn off the stdout reporting for
now.