Support protected branches on GitHub
GitHub recently added a feature called "protected branches", which is
not compatible with the current Homu flow. The main problem is that
there is no way to mark a merge commit that is not yet pushed to a
remote repository as "success".

Until GitHub adds such functionality, work around the issue by pushing
the merge commit to a temporary branch and resetting the target branch
to the merge commit.

Protected branches also conflict with the "fake merge" feature, as it
forbids force-pushing. This cannot be solved mechanically, so inform the
user in such cases.

Fixes #109.